requirements.txt
requests
pydantic
pandas
python-dateutil
pyyaml
reportlab
jinja2
hubspot-api-client
tqdm
